feature(hooks): add custom hooks
- add useDevelopers hooks
- add useRepositories hooks
- modify Developers page to use custom hooks
- modify Repositories page to use custom hooks
