+++ src/pages/Developers/index.tsx
@@ -1,13 +1,11 @@
-import { useState, useEffect, useContext } from "react";
 import { ListGroup } from "react-bootstrap";
-import githubTrends from "github-trends-api";
 
-import { AppContext } from "../../providers/context";
 import Template from "../../components/Template";
 import styles from "./Developers.module.css";
 import DeveloperItem from "../../components/DeveloperItem";
+import useDevelopers from "../../hooks/useDevelopers";
 
-interface IDeveloper {
+export interface IDeveloper {
   author: string;
   username: string;
   avatar: string;
@@ -18,20 +16,7 @@ interface IDeveloper {
 }
 
 const Developers = () => {
-  const { state } = useContext(AppContext);
-  const [developers, setDevelopers] = useState<IDeveloper[]>([]);
-  useEffect(() => {
-    const fetchRepositories = async () => {
-      const data = await githubTrends({
-        section: "developers",
-        language: state?.language?.value || "javascript",
-        since: state?.dateRange?.value,
-      });
-      setDevelopers(data);
-    };
-
-    fetchRepositories();
-  }, [state?.dateRange?.value, state?.language?.value]);
+  const { state, developers } = useDevelopers();
 
   return (
     <Template tab={"developers"}>
