+++ src/pages/Repositories/index.tsx
@@ -1,13 +1,11 @@
-import { useState, useEffect, useContext } from "react";
 import { ListGroup } from "react-bootstrap";
-import githubTrends from "github-trends-api";
 
 import Template from "../../components/Template";
 import RepoItem from "../../components/RepoItem";
-import { AppContext } from "../../providers/context";
 import styles from "./Repositories.module.css";
+import useRepositories from "../../hooks/useRepositories";
 
-interface IRepository {
+export interface IRepository {
   author: string;
   reponame: string;
   repourl: string;
@@ -24,20 +22,7 @@ interface IRepository {
 }
 
 const Repositories = () => {
-  const { state } = useContext(AppContext);
-  const [repos, setRepos] = useState<IRepository[]>([]);
-  useEffect(() => {
-    const fetchRepositories = async () => {
-      const data = await githubTrends({
-        since: state?.dateRange?.value,
-        language: state?.language?.value,
-      });
-      setRepos(data);
-    };
-
-    fetchRepositories();
-  }, [state?.dateRange?.value, state?.language?.value]);
-
+  const { repos, state } = useRepositories();
   return (
     <Template tab="repositories">
       <ListGroup variant="flush">
